requirements.txt
coverage==4.3.4
Django==1.11.1
django-extensions==1.7.8
django-guardian==1.4.8
django-nose==1.4.4
django-npm==1.0.0
factory-boy==2.8.1
